requirements.txt
discord-webhook
praw
python-dotenv